pyproject.toml
[project]
name = "trading"
version = "0.1.0"
description = ""
authors = [
    {name = "novis10813",email = "zhizhongz895253@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp (==3.10.11)",
    "aioredis (==2.0.1)",
    "cython (==3.0.11)",
    "hiredis (==3.1.0)",
    "ipywidgets (==8.1.5)",
    "kafka-python (==2.0.2)",
    "lightgbm (==4.5.0)",
    "llvmlite (==0.43.0)",
    "numba (==0.60.0)",
    "numpy (==1.26.4)",
    "python-dotenv (==1.0.1)",
    "pyyaml (==6.0.2)",
    "redis (==5.2.1)",
    "tqdm (==4.67.1)",
    "uv (==0.5.15)",
    "websockets (==14.1)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
